Separate nturl2path module test from test_urllib.py
