+++ Misc/NEWS.d/next/Tests/2024-07-28-13-41-49.gh-issue-122371.6IYf_p.rst
@@ -1,0 +1,1 @@
+Separate nturl2path module test from test_urllib.py
